feature gated num_traits
